pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evodrive-lab"
version = "0.1.0"
description = "Docker-first driving benchmark with GA, NEAT, and PPO."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Bishal Mahatchhetri" }]
dependencies = [
  "fastapi>=0.115.6,<1.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
  "streamlit>=1.43.0,<2.0.0",
  "requests>=2.32.3,<3.0.0",
  "numpy>=2.1.0,<3.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "sqlmodel>=0.0.22,<0.0.23",
  "gymnasium>=1.0.0,<2.0.0",
  "neat-python>=0.92,<0.93",
  "Box2D>=2.3.2,<2.4.0",
  "pandas>=2.2.0,<3.0.0",
  "plotly>=6.0.0,<7.0.0",
  "matplotlib>=3.9.0,<4.0.0",
  "pygame-ce>=2.5.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-cov>=6.0.0,<7.0.0",
  "httpx>=0.28.0,<0.29.0",
  "ruff>=0.11.0,<0.12.0",
  "mypy>=1.15.0,<2.0.0",
  "pre-commit>=4.1.0,<5.0.0",
]

[project.scripts]
evodrive-api = "app.api.main:run"
evodrive-worker = "app.worker.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
